# backend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Trading demo API",
    "description": "Simulated market data: a fixed instrument universe advancing on a factor model, a library of world events, and a typed contract over both.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/scenario": {
      "get": {
        "tags": [
          "scenario"
        ],
        "summary": "Get Scenario",
        "description": "The active scenario, its headlines, and the tick it was activated at.\n\n`activated_at` is null at baseline. The headlines are served from here rather than\nheld in the client, so a scenario change reaches the ticker without a redeploy.",
        "operationId": "getActiveScenario",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveScenario"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scenario"
        ],
        "summary": "Post Scenario",
        "description": "Make a scenario active from the next tick, stamping `activated_at`.",
        "operationId": "activateScenario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveScenario"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scenario"
        ],
        "summary": "Delete Scenario",
        "description": "Return to baseline, clearing `activated_at`.\n\nThe bars written while the scenario was active stay exactly as they are \u2014 going back\nto normal is a change to what happens next, not an undo.",
        "operationId": "deactivateScenario",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveScenario"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivateScenarioRequest": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ScenarioId",
            "description": "An id the scenario library holds."
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ActivateScenarioRequest",
        "description": "The body of `POST /scenario`. Typed to the enum, which is what makes a 422."
      },
      "ActiveScenario": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "headlines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Headlines",
            "description": "Two or three canned headlines."
          },
          "activated_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At",
            "description": "Tick index the active scenario was activated at; null at baseline. Always present."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "headlines",
          "activated_at"
        ],
        "title": "ActiveScenario",
        "description": "`GET /scenario` \u2014 the active scenario, its headlines and its activation tick."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioId": {
        "type": "string",
        "enum": [
          "baseline",
          "oil_supply_shock",
          "rate_hike_surprise",
          "credit_crunch",
          "risk_rally",
          "dollar_surge",
          "recession_scare"
        ],
        "title": "ScenarioId",
        "description": "The ids in data/scenarios.json, generated from the file at import."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}